Fixing error handling and tests

+++ app/services/db-export/delete-file.service.js
@@ -12,7 +12,9 @@ const fs = require('fs')
  * @param {String} filePath The file path that we want to delete
  */
 async function go (filePath) {
-  fs.unlinkSync(filePath)
+  if (fs.existsSync(filePath)) {
+    fs.unlinkSync(filePath)
+  }
 }
 
 module.exports = {
